+++ bofire/data_models/surrogates/api.py
@@ -1,4 +1,3 @@
-from bofire.data_models.surrogates.bnn import SingleTaskIBNNSurrogate
 from bofire.data_models.surrogates.botorch import BotorchSurrogate
 from bofire.data_models.surrogates.botorch_surrogates import (
     AnyBotorchSurrogate,
@@ -49,7 +48,10 @@ from bofire.data_models.surrogates.single_task_gp import (
 )
 from bofire.data_models.surrogates.surrogate import Surrogate
 from bofire.data_models.surrogates.tanimoto_gp import TanimotoGPSurrogate
-from bofire.data_models.surrogates.trainable_botorch import TrainableBotorchSurrogate
+from bofire.data_models.surrogates.trainable_botorch import (
+    InputScaledBotorchSurrogate,
+    TrainableBotorchSurrogate,
+)
 from bofire.data_models.unions import tagged_union
 
 
@@ -62,13 +64,10 @@ AnySurrogate = tagged_union(
     ClassificationMLPEnsemble,
     RegressionMLPEnsemble,
     FullyBayesianSingleTaskGPSurrogate,
-    LinearSurrogate,
-    PolynomialSurrogate,
     TanimotoGPSurrogate,
     LinearDeterministicSurrogate,
     CategoricalDeterministicSurrogate,
     MultiTaskGPSurrogate,
-    SingleTaskIBNNSurrogate,
     AdditiveMapSaasSingleTaskGPSurrogate,
     EnsembleMapSaasSingleTaskGPSurrogate,
     PairwiseGPSurrogate,
@@ -82,9 +81,6 @@ AnyTrainableSurrogate = tagged_union(
     ClassificationMLPEnsemble,
     RegressionMLPEnsemble,
     FullyBayesianSingleTaskGPSurrogate,
-    LinearSurrogate,
-    PolynomialSurrogate,
-    SingleTaskIBNNSurrogate,
     TanimotoGPSurrogate,
     AdditiveMapSaasSingleTaskGPSurrogate,
     EnsembleMapSaasSingleTaskGPSurrogate,
@@ -98,12 +94,9 @@ AnyRegressionSurrogate = tagged_union(
     MixedSingleTaskGPSurrogate,
     RegressionMLPEnsemble,
     FullyBayesianSingleTaskGPSurrogate,
-    LinearSurrogate,
-    PolynomialSurrogate,
     TanimotoGPSurrogate,
     LinearDeterministicSurrogate,
     MultiTaskGPSurrogate,
-    SingleTaskIBNNSurrogate,
     AdditiveMapSaasSingleTaskGPSurrogate,
     EnsembleMapSaasSingleTaskGPSurrogate,
 )